==> 2-Model.ipynb <==
## Data introduction and import necessary Libraries

**Pixel Resolution:** 30m * 30m<br>
**Target:** Target data set<br>
**Point of Interest (POI):** POI Kernel Density, 
- POI_Sel, BandWidth range: 250 m - 2500 m, interval = 250 m<br>

**Road Network (RN):** Road Kernel Density
- BandWidth range: 250 m - 2500 m, interval = 250 m<br>

**NTL:** Time,2019/03; NPP-VIIRS, DNB<br>
**XM_Boundary:** mask of all the layers<br>
**Train_Test:** 
- Train, data located in Off-island area of Xiamen, label = 0
- Test,  data located in Island area of Xiamen, label = 1

import math
import json
import copy
import time
import numpy as np
import pandas as pd
from sl_1 import * # my custom module 

# Extract and generate headers
# extrat_begin('Target', 'begin')

## read clean data and split it

### read data

clean_data = pd.read_csv(r'data\ly_df_clean.csv')
clean_data.head()

len(clean_data) # Sample size: 1890571

### spli train and test
- Train, data located in Off-island area of Xiamen, label = 0
- Test,  data located in Island area of Xiamen, label = 1

Train = clean_data[clean_data['Train_Test'] == 0]
Test = clean_data[clean_data['Train_Test'] == 1]
# Train.to_csv(r'data\Train.csv')
# Test.to_csv(r'data\Test.csv')
# Train.head()
# Test.head()

len(Test)

## Sensitivity analysis of bandwidth

def get_name(name_pref):
    '''
    input: name_pref, name prefixs of lyer name list
    output: a lyer name list
    '''
    ly_name = [name_pref + str(i) for i in range(250,2750,250)]
    return ly_name
# get_name('poi_all_')  # test example 1
get_name('RN_') # test example 2

from sklearn.model_selection import cross_val_score

def get_name(name_pref):
    '''
    input: name_pref, name prefixs of lyer name list
    output: a lyer name list
    '''
    ly_name = [name_pref + str(i) for i in range(250,2750,250)]
    return ly_name
# get_name('poi_all_')  # test example 1
get_name('RN_') # test example 2
# get_name('poi_all_') + get_name('RN_') + ['Target','Train_Test','XM_Boundary']

def opti_band(names, clf):
    '''input: names, a list made up of different band width name
       output: a string, a band width name with max score
    '''
    score_list = []
    for name in names:
        X = np.array(Train[name]).reshape(-1, 1)
        y = Train['Target']
        scores = cross_val_score(clf, X, y, cv=5)
        score_list.append(scores.mean())
    return score_list

def select_name(names,clf):
    A = {}
    scores = opti_band(names, clf)
    ind = scores.index(max(scores))
    name = names[ind]
    A[name] = scores
    return A
# input_name = 'poi_all_' # 
# names = get_name(input_name)[0:2]
# name = select_name(names)
# name

from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier,GradientBoostingClassifier

names = ["LogisticRegression",
         "Decision Tree", 
         "Random Forest", 
         "AdaBoost",
         "GradientBoosting"]

classifiers = [
    LogisticRegression(random_state=0, solver='lbfgs', max_iter =150),
    DecisionTreeClassifier(random_state=0, max_depth=2),
    RandomForestClassifier(n_estimators=10, random_state=0, max_depth=2),
    AdaBoostClassifier(n_estimators=10, random_state=0),
    GradientBoostingClassifier(n_estimators=100, random_state=0)]

list1 = []
for name, clf in list(zip(names, classifiers))[5:]:
    for input_name in ['poi_sel_', 'RN_']:
        names = get_name(input_name)[0:8]
        name = select_name(names,clf)
        print(name)
        list1.append(name)

score_list = [{'poi_sel_750': [0.8801062160559965,
   0.8867527770787383,
   0.8871036003829122,
   0.88501723456701,
   0.8810242833197899,
   0.878080327222498,
   0.8766112574295093,
   0.8748780477656201]},
 {'RN_1750': [0.8407899344127374,
   0.8488049048455742,
   0.8535896834056246,
   0.8559721743507032,
   0.856960657622183,
   0.8573423179653057,
   0.8577001285138349,
   0.8572247863462765]},
 {'poi_sel_500': [0.8780075974432119,
   0.9035198512662044,
   0.8835318123221562,
   0.8625925726315403,
   0.8516888788466208,
   0.8693148178666711,
   0.8469715763557708,
   0.8376492541726485]},
 {'RN_750': [0.8460674913289254,
   0.8527024349863552,
   0.8588882041349454,
   0.846455513089673,
   0.8438501720565036,
   0.8429728054711323,
   0.8449323214324895,
   0.8581172937545791]},
 {'poi_sel_500': [0.8786912204900327,
   0.9036239949351265,
   0.8849328071571725,
   0.8671527731376203,
   0.8529327826371931,
   0.8689610771140028,
   0.8479333041995387,
   0.8418842339979002]},
 {'RN_750': [0.8453786294051039,
   0.8524132746047602,
   0.8594868808268341,
   0.8464223502623358,
   0.8451126929524561,
   0.8434830499358881,
   0.8455600924692419,
   0.8538759009698118]},
 {'poi_sel_750': [0.8875387488088566,
   0.890418669984753,
   0.8918958601815452,
   0.8893609002583232,
   0.8668036183400524,
   0.8549993470831492,
   0.8625530037806393,
   0.8576501071542906]},
 {'RN_250': [0.8350212943631554,
   0.8296034518019795,
   0.8306128904054318,
   0.8288098146822026,
   0.8292048688247233,
   0.8275618438799466,
   0.8248471098005687,
   0.8234868432588511]},
 {'poi_sel_500': [0.8883125815544191,
   0.8982155132961005,
   0.8930321690121877,
   0.8854570172655682,
   0.8794562478978378,
   0.8797227158978685,
   0.872141768356191,
   0.8647242889915333]},
 {'RN_1250': [0.8396937826843999,
   0.8455211435736503,
   0.8511629173226114,
   0.8537013465738028,
   0.8548754306914967,
   0.8526756166296728,
   0.8524516187250724,
   0.8503646761465044]}]

# score_list

## Ensemble methods
- RandomForest
- AdaBoost
- Gradient Tree Boosting
- Voting Classifier

### cross validation

import copy
import numpy as np
from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier,GradientBoostingClassifier
from sklearn.metrics import f1_score
from sklearn.model_selection import GridSearchCV

names = ["LogisticRegression",
         "Naive Bayes", 
         "QDA",
         "Decision Tree", 
         "Random Forest", 
         "AdaBoost",
         "GradientBoosting"]

params = [{'param': 'default'},
          {'param': 'default'},
          {'param': 'default'},
          {'max_depth': list(range(1, 6))},
          {'max_depth': list(range(1, 6))},
          {'learning_rate': [0.001,0.1,1]}, # 
          {'learning_rate': [0.001,0.1,1]}]

classifiers = [
    LogisticRegression(random_state=0, solver='lbfgs', max_iter =150),
    GaussianNB(),
    QuadraticDiscriminantAnalysis(),
    DecisionTreeClassifier(random_state=0),
    RandomForestClassifier(n_estimators=100, random_state=0),
    AdaBoostClassifier(n_estimators=100, random_state=0),
    GradientBoostingClassifier(n_estimators=100, random_state=0)]


X_Train =  np.array(Train[['poi_sel_250', 'RN_500','NTL']])
Y_Train = Train['Target'].copy()
X_Test = np.array(Test[['poi_sel_250', 'RN_500','NTL']])
Y_Test = Test['Target'].copy() # get sub dataframe

params_cv = [] # params selcted by cross validation
for name, params, clf in list(zip(names, params, classifiers)):
    print(name)
    # for "Decision Tree", "Random Forest", "AdaBoost", "GradientBoosting", conduct cross validation.
    if name in [ "Decision Tree", "Random Forest", "AdaBoost", "GradientBoosting"]:
        grid = GridSearchCV(clf, params, cv=10, iid=False)
        grid.fit(X_Train, Y_Train)
        # use the best estimator to compute the kernel density estimate
        clf_best = grid.best_estimator_
        if name in [ "Decision Tree", "Random Forest"]:
            params_cv.append(clf_best.max_depth)
        else:
            params_cv.append(clf_best.learning_rate)
    else:
        pass

params_cv

### example 1 

import warnings
warnings.filterwarnings("ignore")

import copy
import numpy as np
from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier,GradientBoostingClassifier
from sklearn.metrics import f1_score
from sklearn.model_selection import GridSearchCV

names = ["LogisticRegression",
         "Naive Bayes", 
         "QDA",
         "Decision Tree", 
         "Random Forest", 
         "AdaBoost",
         "GradientBoosting"]

classifiers = [
    LogisticRegression(random_state=0, solver='lbfgs', max_iter =150),
    GaussianNB(),
    QuadraticDiscriminantAnalysis(),
    DecisionTreeClassifier(random_state=0, max_depth=2),
    RandomForestClassifier(n_estimators=100, random_state=0, max_depth=4),
    AdaBoostClassifier(n_estimators=100, random_state=0, learning_rate=1),
    GradientBoostingClassifier(n_estimators=100, random_state=0, learning_rate=0.1)]

X_Train =  np.array(Train[['poi_sel_250', 'RN_500','NTL']])
Y_Train = Train['Target'].copy()
X_Test = np.array(Test[['poi_sel_250', 'RN_500','NTL']])
Y_Test = Test['Target'].copy() # get sub dataframe

for name, clf in list(zip(names, classifiers)):
    clf.fit(X_Train, Y_Train)
    score = clf.score(X_Test, Y_Test) # overall_accuracy
    predictions = clf.predict(X_Test)
    f1 = f1_score(Y_Test, predictions, average='macro') # f1_score
    print('name',name,'overall_accuracy:', score,'f1_score',f1)

### example2-reverse Train and test

X_Train =  np.array(Train[['poi_sel_250', 'RN_500','NTL']])
Y_Train = Train['Target'].copy()
X_Test = np.array(Test[['poi_sel_250', 'RN_500','NTL']])
Y_Test = Test['Target'].copy() # get sub dataframe

for name, clf in list(zip(names, classifiers)):
    clf.fit(X_Test, Y_Test)
    score = clf.score(X_Train, Y_Train) # overall_accuracy
    predictions = clf.predict(X_Train)
    f1 = f1_score(Y_Train, predictions, average='macro') # f1_score
    print('name',name,'overall_accuracy:', score,'f1_score',f1)

### write to CSV

import copy
import numpy as np
from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier,GradientBoostingClassifier
from sklearn.metrics import f1_score
from itertools import combinations
def get_comb(input_list,n):
    '''input: input_list, a list.
       input: n，an integer; the number of combinations.
       output: a list, the combinations of input_list.
    '''
    comb_all = []
    m = 1
    while m<n+1:
        a = [list(i) for i in list(combinations(input_list, m))]
        comb_all += a
        m += 1
    return comb_all
# comb_all = get_comb([1,2,3], 3)
# comb_all

comb_all = get_comb(['poi_sel_250', 'RN_500','NTL'], 3)
comb_all

def clf_pre(data_name,names, classifiers):
    '''input: data_name,a list consist of the names or name of data to be predicted by different classfiers. 
       names: the name of different classifiers.
       classifiers: the classifiers to predict data.
    '''
    X_Train =  np.array(Train[data_name])
    Y_Train = Train['Target'].copy()
    X_Test = np.array(Test[data_name])
    Y_Test = Test['Target'].copy() # get sub dataframe

    score_list = []
    f1_list = []
    pred = []
    pred_proba = []
    for name, clf in list(zip(names, classifiers)):
        clf.fit(X_Train, Y_Train)
        score = clf.score(X_Test, Y_Test) # overall_accuracy
        predictions = clf.predict(X_Test)
        proba = clf.predict_proba(X_Test)
        f1 = f1_score(Y_Test, predictions, average='macro') # f1_score
        score_list.append(score)
        f1_list.append(f1)
        pred.append(predictions)
        pred_proba.append(proba)
        if len(data_name) > 0:
            dname = '_'.join(data_name)
        else:
            dname = data_name[0]
        print('data_name', dname, 'clf_name', name,'overall_accuracy:', score,'f1_score',f1)
    data1 = {'data_name': dname, 'clf_name': names, 
             'accuracy': score_list, 'f1_score': f1_list, 
             'pred': pred, 'pred_proba': pred_proba}
    DF1 = pd.DataFrame(data1)
    return DF1

def clf_pres(data_names,names, classifiers):
    '''input: data_name,a list consist of the names or name of data to be predicted by different classfiers. 
       names: the name of different classifiers.
       classifiers: the classifiers to predict data.
    '''
    list1 = []
    for data_name in data_names:
        DF1 = clf_pre(data_name,names, classifiers)
        list1.append(DF1)
    DF2 = pd.concat(list1)
    return DF2


data_names = get_comb(['poi_sel_250', 'RN_500','NTL'], 3)

names = ["LogisticRegression",
#          "Naive Bayes", 
#          "QDA",
         "Decision Tree", 
         "Random Forest", 
         "AdaBoost",
         "GradientBoosting"
        ]

classifiers = [
    LogisticRegression(random_state=0, solver='lbfgs', max_iter =150),
#     GaussianNB(),
#     QuadraticDiscriminantAnalysis(),
    DecisionTreeClassifier(random_state=0, max_depth=2),
    RandomForestClassifier(n_estimators=100, random_state=0, max_depth=4),
    AdaBoostClassifier(n_estimators=100, random_state=0, learning_rate=1),
    GradientBoostingClassifier(n_estimators=100, random_state=0, learning_rate=0.1)
]

DF = clf_pres(data_names,names, classifiers)

DF.head()

np.vstack(DF['pred']).T.shape

pred = pd.DataFrame(np.vstack(DF['pred']).T)
pred.head()

pred_proba_0 = pd.DataFrame(np.vstack([i[:,0] for i in np.array(DF['pred_proba'])]).T)
pred_proba_0.head()

pred_proba_1 = pd.DataFrame(np.vstack([i[:,1] for i in np.array(DF['pred_proba'])]).T)
pred_proba_1.head()

DF.to_csv("data/pred_OA_F1.csv",index=False)
pred.to_csv("data/pred_test.csv",index=False)
pred_proba_0.to_csv("data/pred_proba0_test.csv",index=False)
pred_proba_1.to_csv("data/pred_proba1_test.csv",index=False)

## AUC

from sklearn.metrics import roc_auc_score

clean_data = pd.read_csv(r'data\ly_df_clean.csv')

Test = clean_data[clean_data['Train_Test'] == 1]['Target']

proba1 = pd.read_csv(r"data\pred_proba1_test.csv")

Test.head()

proba1.head()

y_true = Test
auc_list = []
for col in range(35):
    y_scores = proba1.iloc[:,col]
    auc = roc_auc_score(y_true, y_scores)
    list1.append(auc)

auc_list

auc = pd.DataFrame(auc_list, columns=["auc"])

auc.to_csv('data/AUC.csv',index=False)
