==> 3-data_to_ASCII.ipynb <==
## Data introduction and import necessary Libraries

**Pixel Resolution:** 30m * 30m<br>
**Target:** Target data set<br>
**Point of Interest (POI):** POI Kernel Density, 
- POI_All, BandWidth range: 250 m - 2500 m, interval = 250 m<br>
- POI_Sel, BandWidth range: 250 m - 2500 m, interval = 250 m<br>
- POI_All: 1000m; POI_Sel: 500m<br>

**Road Network (RN):** Road Kernel Density
- BandWidth range: 250 m - 2500 m, interval = 250 m<br>
- BandWidth: 1250m<br>

**NTL:** Time,2019/03; NPP-VIIRS, DNB<br>
**XM_Boundary:** mask of all the layers<br>
**Train_Test:** 
- Train, data located in Off-island area of Xiamen, label = 0
- Test,  data located in Island area of Xiamen, label = 1

import math
import json
import copy
import time
import numpy as np
np.set_printoptions(suppress=True)
import pandas as pd
from sl_1 import * # my custom module 

# Extract and generate headers
# extrat_begin('Target', 'begin')

## read clean data and split it

clean_data = pd.read_csv(r'data\ly_df_clean.csv')

clean_new = copy.deepcopy(clean_data[['ID','Target','Train_Test']])
clean_new.head()

len(clean_new) # Sample size: 1890571

Train = clean_data[clean_data['Train_Test'] == 0]
Test = clean_data[clean_data['Train_Test'] == 1]

##  convert pred data

# pred_df = pd.read_csv("data/pred_df.csv")
# # pred_new = copy.deepcopy(pred_df[~pred_df['clf_name'].isin(['Naive Bayes', 'QDA'])])
# pred_new['ID'] = list(range(0,len(pred_new)))
# pred_new.head()

pred_test = pd.read_csv("data/pred_test.csv")

def function1(a, b): # a: true target; b: predict
    '''
    a, One-dimensional array
    b，One-dimensional array
    '''
    arr1 = np.zeros(shape=(len(a),1))
    arr1[(b == 0) & (a == 0)] = 0 # TN
    arr1[(b == 1) & (a == 1)] = 1 # TP
    arr1[(b == 1) & (a == 0)] = 2 # FP
    arr1[(b == 0) & (a == 1)] = 3 # FN
    return arr1
# a = np.array([0,0,1,1])
# b = np.array([0,1,1,0])
# function1(a, b)

def function2(a, c): # a: true target; c: predict
    '''
    a, One-dimensional array
    c，n-dimensional array
    '''
    arr2 = np.zeros(c.shape)
    for i in range(c.shape[1]):
        b = c[:,i]
#         print('b',b)
        arr2[:,i] = function1(a, b).reshape(-1)
    return arr2

a = np.array(Test.Target)
c = np.array(pred_test)
d = function2(a, c)

d.shape

pred_test1 = pd.DataFrame(d)

## Genetate Index

### index test

pred_test1['ID'] = list(clean_new.iloc[:,0][clean_new['Train_Test'] == 1]) # Test data
pred_test1.head()

### index train

Train_1 = np.array([Train['Target']] * 35).T

Train_1.shape

Train_2 = pd.DataFrame(Train_1)
Train_2['ID'] = list(clean_new.iloc[:,0][clean_new['Train_Test'] == 0]) # Train data
Train_2.head()

### merge Train and test

Train_Test_arr = np.vstack([np.array(pred_test1).astype(np.int64),np.array(Train_2).astype(np.int64)])

Train_Test = pd.DataFrame(Train_Test_arr,columns= list(range(35)) + ['ID'])

Train_Test = Train_Test.sort_values(by="ID" , ascending=True)

Train_Test.head()

### rebuild index

Raw_index = pd.DataFrame(list(range(3450552)),columns=['ID'])

data1 = Train_Test.copy()

data2 = pd.merge(Raw_index, data1, how='left', left_on='ID', right_on='ID', sort=True)

data3 = data2.copy().fillna(-9999)

data3.head()

shape_ras = (1786, 1932) # Number of rows and columns
arr3 = np.array(data3).astype(np.int64)
j = arr3.shape[1]
for i in range(1,j):
    pred_y1 = arr3[:,i]
    
    pred_y2 = pred_y1.reshape(shape_ras)
    np.savetxt('data/pred_%s.txt' % (i), pred_y2, fmt='%0.0f') # Skip the first 6 lines
    add_begin('pred_%s' % (i))

## to ArcGIS Pro

# -*- coding: utf-8 -*-

import arcpy

# To allow overwriting the outputs change the overwrite option to true.
arcpy.env.overwriteOutput = True

# Local variables:
for i in range(1,36):
    pred_1_txt = r"E:\Paper\XM_Urban_Built\Data\pred_35\pred_%s.txt" % (i)
    pred_1 = r"E:\Paper\XM_Urban_Built\XM_Urban_Built.gdb\pred_%s" % (i)
    Update_Input_Dataset_or_Feature_Class = pred_1

    # Process: ASCII to Raster
    arcpy.ASCIIToRaster_conversion(in_ascii_file=pred_1_txt, out_raster=pred_1, data_type="INTEGER")

    # Process: Define Projection
    arcpy.DefineProjection_management(in_dataset=pred_1, coor_system="PROJCS['UTM_Zone_50N',GEOGCS['GCS_WGS_1984',DATUM['D_WGS_1984',SPHEROID['WGS_1984',6378137.0,298.257223563]],PRIMEM['Greenwich',0.0],UNIT['Degree',0.0174532925199433]],PROJECTION['Transverse_Mercator'],PARAMETER['false_easting',500000.0],PARAMETER['false_northing',0.0],PARAMETER['central_meridian',117.0],PARAMETER['scale_factor',0.9996],PARAMETER['latitude_of_origin',0.0],UNIT['Meter',1.0]]")
